Split answer DTOs into request and response subfolders
